Use ScrollToButton instead of floating form to notify user there's an offer near top of the page
