--- apps/store/src/components/ProductPage/PriceFooter/PriceFooter.stories.tsx
@@ -1,24 +1,0 @@
-import { ComponentMeta, Story } from '@storybook/react'
-import * as PriceFooter from './PriceFooter'
-
-export default {
-  title: 'Product Page / Price Footer',
-  component: PriceFooter.Footer,
-  argTypes: {
-    onClick: { action: 'clicked' },
-  },
-} as ComponentMeta<typeof PriceFooter.Footer>
-
-const Template: Story<PriceFooter.ButtonProps> = (props) => {
-  return (
-    <PriceFooter.Footer {...props}>
-      <PriceFooter.Button onClick={props.onClick} disabled={props.disabled}>
-        <span>SEK 149/mp.</span>
-        <span>Add to cart</span>
-      </PriceFooter.Button>
-    </PriceFooter.Footer>
-  )
-}
-
-export const Default = Template.bind({})
-Default.args = {}
