+++ apps/store/src/components/ProductPage/ScrollToButton/ScrollToButton.tsx
@@ -1,11 +1,12 @@
 import styled from '@emotion/styled'
+import { ReactNode } from 'react'
 import { ChevronIcon } from 'ui'
 import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
 
 export type ScrollToButtonProps = {
   type: 'button' | 'submit'
   targetRef: React.RefObject<HTMLElement>
-  children: string
+  children: ReactNode
 }
 
 export const ScrollToButton = ({ children, type, targetRef }: ScrollToButtonProps) => {
@@ -15,7 +16,7 @@ export const ScrollToButton = ({ children, type, targetRef }: ScrollToButtonProp
     <Wrappper>
       <StyledPillButton type={type} onClick={handleClick}>
         <SpaceFlex space={0.5} align="center">
-          <TextLabel>{children}</TextLabel>
+          {children}
           <ChevronUp size="0.75rem" />
         </SpaceFlex>
       </StyledPillButton>
@@ -31,17 +32,12 @@ const Wrappper = styled.div(({ theme }) => ({
 }))
 
 const StyledPillButton = styled.button(({ theme }) => ({
+  color: theme.colors.white,
   height: '2rem',
   borderRadius: '1rem',
   backgroundColor: theme.colors.black,
   padding: `0 ${theme.space[4]}`,
   cursor: 'pointer',
-}))
-
-const TextLabel = styled.span(({ theme }) => ({
-  marginTop: -2,
-  color: theme.colors.white,
-  lineHeight: 1,
 }))
 
 const ChevronUp = styled(ChevronIcon)(({ theme }) => ({
